### 2019-10-07 #1789 @whedon
[PRE REVIEW]: SunPy: A Python package for Solar Physics
**Submitting author:** @Cadair (<a href="http://orcid.org/0000-0003-4217-4642">Stuart Mumford</a>)
**Repository:** <a href="https://github.com/sunpy/sunpy" target ="_blank">https://github.com/sunpy/sunpy</a>
**Version:** v1.0.0
