### 2019-10-07 #1789 @arfon
:wave: @Cadair - thanks for your submission to JOSS. Does this JOSS submission accompany a publication in a AAS journal?
