### 2019-10-07 #1789 @ehsteve
@arfon FYI, there is a companion paper that was submitted and accepted by ApJ.
