### 2019-10-07 #1789 @arfon
> @arfon FYI, there is a companion paper that was submitted and accepted by ApJ. I am the corresponding author for that paper.

/ cc @crawfordsm
