### 2019-10-07 #1789 @arfon
@Cadair [this list of people](https://bit.ly/joss-reviewers) posted at the top of this thread have volunteered but feel free to suggest people from outside of this list who you think could give a good review.
