### 2019-10-07 #1789 @arfon
Also, given this is a joint submission with AAS Journals, could you add metadata like this to the YAML header of the `paper.md` file @Cadair?

```
aas-doi: DOI of AAS paper
aas-journal: Name of journal (e.g. Astronomical Journal)
```
