### 2019-10-07 401deb4
Resolve one deprecation warning from test_sysinfo.py
xref #2886
using the module `distro` for determining the linux_distribution instead of `platform`.
