### 2019-10-08 0f1830e
added a dict with a new key and value, then updated the dict with kwargs
