+++ changelog/3397.feature.rst
@@ -1,0 +1,1 @@
+Added a new key - annotation to kwergs of map.draw_grid
