### 2019-10-08 965dbf8
Update sunpy/map/mapbase.py
Co-Authored-By: Nabil Freij <nabil.freij@gmail.com>
