### 2019-10-08 ccde8c6
Update changelog/3397.feature.rst
Co-Authored-By: Nabil Freij <nabil.freij@gmail.com>
