+++ changelog/3397.feature.rst
@@ -1,1 +1,1 @@
-Added a new key - annotation to kwergs of map.draw_grid
+Now able to turn off the tick and axes labels (``annotate=False``) to `sunpy.map.GenericMap.draw_grid`.
