### 2019-10-08 75de245
Update sunpy/visualization/wcsaxes_compat.py
Co-Authored-By: Nabil Freij <nabil.freij@gmail.com>
