### 2019-10-08 29477a7
Apply suggestions from code review
Co-Authored-By: Albert Y. Shih <ayshih@gmail.com>
