+++ changelog/3360.feature.rst
@@ -1,2 +1,2 @@
-Added the `sunpy.coordinates.sun.carrington_rotation_starttime` function to
-compute the start time of a given Carrington rotation number.
+Added the `sunpy.coordinates.sun.carrington_rotation_time` function to
+compute the time of a given Carrington rotation number.
