### 2019-10-08 26db0f5
Apply suggestions from code review
Co-Authored-By: Albert Y. Shih <ayshih@gmail.com>
