+++ sunpy/coordinates/tests/test_sun.py
@@ -263,8 +263,7 @@ def test_carrington_rotation_number(date, day_fraction, rotation_number):
                           ])
 def test_carrington_rotation_starttime(crot, julian_days):
     # Stated precision in the docstring is 0.11 seconds
-    atol = 0.11 / (60 * 60 * 24)
-    assert_allclose(sun.carrington_rotation_time(crot).tt.jd, julian_days, atol=atol)
+    assert_quantity_allclose(sun.carrington_rotation_time(crot).tt.jd * u.day, julian_days * u.day, atol=0.11*u.s)
 
 
 def test_carrington_rotation_roundtrip():
@@ -272,6 +271,5 @@ def test_carrington_rotation_roundtrip():
     crot = sun.carrington_rotation_number(t)
     t_roundtrip = sun.carrington_rotation_time(crot)
     dt = t - t_roundtrip
-    dt.format = 'sec'
     # Stated precision in the docstring is 0.11 seconds
-    assert np.abs(dt.value) < 0.11
+    assert_quantity_allclose(dt.to(u.s), 0*u.s, atol=0.11*u.s)
