### 2019-10-08 a6d33f5
chore: review comments
remove used __init__, use sunpy instead of sys in conftest
