+++ sunpy/data/__init__.py
@@ -1,7 +1,6 @@
-import sys
-
 import astropy.units as u
 
+import sunpy
 from sunpy import config
 from sunpy.data._sample import download_sample_data
 from sunpy.data.data_manager.cache import Cache
@@ -13,7 +12,7 @@ from sunpy.util.config import CACHE_DIR
 _download_dir = config.get('downloads', 'remote_data_manager_dir')
 
 
-if hasattr(sys, '_called_from_test'):
+if hasattr(sunpy, '_called_from_test'):
     # sunpy executed with pytest
     # this will replace the real cache with a mock cache
     # this is needed for tests
