### 2019-10-08 fdc0316
chore: review comments
remove used __init__, use sunpy instead of sys in conftest
