### 2019-10-08 8c281da
Apply suggestions from code review
Co-Authored-By: Albert Y. Shih <ayshih@gmail.com>
