### 2019-10-08 5148652
Copy the library `distro` into `sunpy/extern`
This replaces the deprecated `platform.linux_distribution`
