### 2019-10-08 6a31428
fix(data manager): do not raise error when the file has been tampered
