### 2019-10-08 01b1af5
fix(data manager): do not raise error when the file has been tampered
