### 2019-10-09 eb16126
Added ASDF schemas for new coordinate frames
