+++ sunpy/io/special/asdf/schemas/sunpy.org/sunpy/coordinates/frames/heliocentricinertial-1.0.0.yaml
@@ -1,0 +1,31 @@
+%YAML 1.1
+---
+$schema: "http://stsci.edu/schemas/yaml-schema/draft-01"
+id: "http://sunpy.org/schemas/sunpy/coordinates/frames/heliocentricinertial-1.0.0"
+tag: "tag:sunpy.org:sunpy/coordinates/frames/heliocentricinertial-1.0.0"
+
+title: |
+  Represents a Heliocentric Inertial (HCI) coordinate object from SunPy
+
+examples:
+  -
+    - A HeliocentricInertial frame without data
+    - |
+        !<tag:sunpy.org:sunpy/coordinates/frames/heliocentricinertial-1.0.0>
+          frame_attributes: {}
+  -
+    - A HeliocentricInertial frame with an obstime
+    - |
+        !<tag:sunpy.org:sunpy/coordinates/frames/heliocentricinertial-1.0.0>
+          frame_attributes: {obstime: !time/time-1.1.0 '2011-01-01T01:01:01.000'}
+
+allOf:
+  - $ref: "tag:astropy.org:astropy/coordinates/frames/baseframe-1.0.0"
+  - properties:
+      frame_attributes:
+        type: object
+        properties:
+          obstime:
+            $ref: "tag:stsci.edu:asdf/time/time-1.1.0"
+        additionalProperties: False
+...
