### 2019-10-10 #1789 @Cadair
maybe (without mentioning them) mwcraig ceb8 or dhuppenkothen

I will add the aas metadata (although I don't know the DOI)
