### 2019-10-10 dcd456f
Add first round of affiliations and orcids
Co-Authored-By: Nabil Freij <nabil.freij@gmail.com>
Co-Authored-By: Steven Christe <ehsteve@users.noreply.github.com>
Co-Authored-By: Albert Y. Shih <ayshih@gmail.com>
Co-Authored-By: David Pérez-Suárez <dps.helio@gmail.com>
Co-Authored-By: aringlis <a.r.inglis@gmail.com>
Co-Authored-By: Brigitta Sipocz <bsipocz@gmail.com>
Co-Authored-By: Drew Leonard <andy.j.leonard@gmail.com>
Co-Authored-By: David Stansby <dstansby@gmail.com>
Co-Authored-By: Laura Hayes <hayesla@tcd.ie>
Co-Authored-By: Nitin Choudhary <nitin.iitkgp23@gmail.com>
Co-Authored-By: Prate
[…]
