+++ joss_paper/paper.md
@@ -9,32 +9,57 @@ authors:
     orcid: 0000-0003-4217-4642
     affiliation: "1, 2"
   - name: Nabil Freij
+    orcid: 0000-0002-6253-082X
+    affiliation: "Institute for Environmental Analytics, University of Reading, Reading RG6 6BX"
   - name: Steven Christe
+    orcid: 0000-0001-6127-795X
+    affiliation: "NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA"
   - name: Jack Ireland
   - name: Florian Mayer
   - name: Keith Hughitt
   - name: Albert Y. Shih
+    orcid: 0000-0001-6874-2594
+    affiliation: "NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA"
   - name: Daniel Ryan
   - name: Simon Liedtke
   - name: David Pérez-Suárez
+    orcid: 0000-0003-0784-6909
+    affiliation: "University College London, Gower Street, London, UK"
   - name: Pritish Chakraborty
   - name: Vishnunarayan K I.
   - name: Andrew Inglis
+    orcid: 0000-0003-0656-2437
+    affiliation: "Catholic University of America / NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA"
   - name: Punyaslok Pattnaik
-  - name: Brigitta Sipocz
+  - name: Brigitta Sipőcz
+    orcid: 0000-0002-3713-6337
+    affiliation: "DIRAC Institute, Department of Astronomy, University of Washington, Seattle, WA 98195, USA"
   - name: Rishabh Sharma
   - name: Andrew Leonard
+    orcid: 0000-0001-5270-7487
+    affiliation: "Aperio Software Ltd., Headingley Enterprise and Arts Centre, Bennett Road, Leeds LS6 3HN"
   - name: David Stansby
+    orcid: 0000-0002-1365-1908
+    affiliation: 'Mullard Space Science Laboratory, University College London, Surrey, UK'
   - name: Russell Hewett
   - name: Alex Hamilton
   - name: Laura Hayes
+    orcid: 0000-0002-6835-2390
+    affiliation: 'NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA'
   - name: Asish Panda
   - name: Matt Earnshaw
   - name: Nitin Choudhary
+    orcid : 0000-0001-6915-4583
+    affiliation: 'Department of Mathematics, Indian Institute of Technology, Kharagpur, 721302, India'
   - name: Ankit Kumar
   - name: Prateek Chanda
+    orcid: 0000-0002-7068-2866
+    affiliation: 'Department of Computer Science & Technology, Indian Institute of Engineering Science & Technology, Shibpur, 711103, India'
   - name: Md Akramul Haque
+    affiliation: 'Department of Mechanical Engineering, ZHCET, Aligarh Muslim University, Aligarh, India'
   - name: Michael S Kirk
+    orcid: 0000-0001-9874-1429
+    affiliation: 'Catholic University of America / NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA'
   - name: Michael Mueller
   - name: Sudarshan Konge
   - name: Rajul Srivastava
@@ -42,26 +67,40 @@ authors:
   - name: Samuel Bennett
   - name: Ankit Baruah
   - name: Will Barnes
+    orcid: 0000-0001-9642-6089
+    affiliation: Lockheed Martin Solar and Astrophysics Laboratory / Bay Area Environmental Research Institute
   - name: Michael Charlton
   - name: Cubostar
   - name: Shane Maloney
+    orcid: 0000-0002-4715-1805
[…]
