### 2019-10-10 0d23323
Extra metadata fields and standardise affiliations
