+++ joss_paper/paper.md
@@ -202,6 +202,8 @@ authors:
     affiliation: International Institute of Information Technology, Hyderabad
   - name: Sarthak Jain
   - name: Tannmay Yadav
+    orcid: 0000-0002-3143-5635
+    affiliation: "Department of Chemical Engineering, Indian Institute of Technology Kharagpur, 721302, India"
   - name: Tessa D. Wilkinson
   - name: Tiago M. D. Pereira
     orcid: 000-0003-4747-4329
