### 2019-10-10 f9fbaa9
Removing .fits from kwargs path
