### 2019-10-10 #1789 @arfon
@ceb8 @dhuppenkothen - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
