### 2019-10-11 cd79726
Add Dan Ryan to author list
