+++ changelog/3399.doc .rst
@@ -1,0 +1,1 @@
+Removing `.fits` from the end of path kwargs in `sunpy.net.FIDO.fetch` docs to change output file extension from `{file}.fits.fits` to `{file}.fits`.
