### 2019-10-12 218e16a
Delay all matplotlib imports to inside plot_settings and plot methods
