+++ sunpy/map/sources/suvi.py
@@ -1,17 +1,6 @@
 """SUVI Map subclass definitions"""
-from __future__ import absolute_import, print_function, division
-
-# pylint: disable=W0221,W0222,E1101,E1121
-
-__author__ = "Jack Ireland"
-__email__ = "jack.ireland@nasa.gov"
-
-import matplotlib.pyplot as plt
-
-from astropy.coordinates import CartesianRepresentation
 import astropy.units as u
-from astropy.visualization.mpl_normalize import ImageNormalize
-from astropy.visualization import AsinhStretch
+from astropy.coordinates import CartesianRepresentation
 
 from sunpy.map import GenericMap
 from sunpy.map.sources.source_type import source_stretch
@@ -81,10 +70,20 @@ class SUVIMap(GenericMap):
         self.meta["detector"] = "SUVI"
         self.meta["telescop"] = "GOES-R"
         self._nickname = self.detector
-        self.plot_settings["cmap"] = plt.get_cmap(self._get_cmap_name())
-        self.plot_settings["norm"] = ImageNormalize(
+
+    def _default_plot_settings(self):
+        import matplotlib.pyplot as plt
+
+        from astropy.visualization import AsinhStretch
+        from astropy.visualization.mpl_normalize import ImageNormalize
+
+        plot_settings = super()._default_plot_settings()
+        plot_settings["cmap"] = plt.get_cmap(self._get_cmap_name())
+        plot_settings["norm"] = ImageNormalize(
             stretch=source_stretch(self.meta, AsinhStretch(0.01))
         )
+
+        return plot_settings
 
     @property
     def _supported_observer_coordinates(self):
