+++ sunpy/map/sources/iris.py
@@ -40,7 +40,8 @@ class SJIMap(GenericMap):
         # Assume pixel units are arcesc if not given
         header['cunit1'] = header.get('cunit1', 'arcsec')
         header['cunit2'] = header.get('cunit2', 'arcsec')
-        GenericMap.__init__(self, data, header, **kwargs)
+
+        super().__init__(data, header, **kwargs)
 
         self.meta['detector'] = "SJI"
         self.meta['waveunit'] = "Angstrom"

[…]
