### 2019-10-13 071f4a7
Fix markup in Map example gallery README
