+++ examples/map/README.txt
@@ -1,4 +1,5 @@
 Map
 ===
 
-This section contains any examples which showcase how SunPy's ~`sunpy.map.Map` can be used to with solar data.
+This section contains any examples which showcase how SunPy's :class:`~sunpy.map.Map`
+can be used to with solar data.
