### 2019-10-13 1e341c3
Add Carrington reprojection example
