### 2019-10-13 0a4c056
Add Carrington reprojection example
