+++ tox.ini
@@ -41,6 +41,7 @@ deps =
     astropy
     astroquery
     jplephem
+    reproject
 commands = sphinx-build docs docs/_build/html -W -b html -d docs/_build/.doctrees
 
 # This env requires tox-conda.
