### 2019-10-14 69f6ed8
Convert axis_ranges for slider axes to callables.
All the schematics of axis_ranges stays the same, but it now also allows
callables for slider axes, and all values for slider axes are converted to a callable.
