### 2019-10-14 9fc58b1
sort lists of maps from glob and directory
