### 2019-10-14 872fac9
test that glob and dir map lists are sorted
