### 2019-10-14 34106aa
Implement a new 1D or 2D capable version of ArrayAnimator using WCS
This is intended to replace ImageAnimatorWCS, and to provide the functionality
of LineAnimator when using WCSAxes.
