+++ sunpy/visualization/animator/base.py
@@ -8,7 +8,7 @@ import mpl_toolkits.axes_grid1.axes_size as Size
 import numpy as np
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 
-__all__ = ['BaseFuncAnimator', 'ArrayAnimator']
+__all__ = ['BaseFuncAnimator', 'ArrayAnimator', 'ArrayAnimatorWCS']
 
 
 class BaseFuncAnimator:
@@ -590,6 +590,127 @@ class ArrayAnimator(BaseFuncAnimator, metaclass=abc.ABCMeta):
         # Update slider label to reflect real world values in axis_ranges.
         label = self.axis_ranges[ax_ind](ind)
         slider.valtext.set_text(f"{label}")
+
+
+from astropy.wcs.wcsapi import BaseLowLevelWCS, SlicedLowLevelWCS
+
+
+class ArrayAnimatorWCS(ArrayAnimator):
+    """
+    Plot a 1 or 2D slice through a ND array with a WCS.
+    """
+
+    def __init__(self, data, wcs, slices, coord_params=None, **kwargs):
+        if not isinstance(wcs, BaseLowLevelWCS):
+            raise ValueError("A WCS object should be provided that implements the astropy WCS API.")
+        if wcs.pixel_n_dim is not data.ndim:
+            raise ValueError("Dimensionality of the data and WCS object do not match.")
+        if len(slices) != wcs.pixel_n_dim:
+            raise ValueError("slices should be the same length as the number of pixel dimensions.")
+        if "x" not in slices:
+            raise ValueError("slices should contain at least 'x' to indicate the axis to plot on the x axis.")
+
+        self.plot_dimensionality = 1
+
+        image_axes = [slices[::-1].index("x")]
+        if "y" in slices:
+            image_axes.append(slices[::-1].index("y"))
+            self.plot_dimensionality = 2
+
+        self.naxis = data.ndim
+        self.num_sliders = self.naxis - self.plot_dimensionality
+        self.slices_wcsaxes = list(slices)
+        self.wcs = wcs
+        self.coord_params = coord_params
+
+        super().__init__(data, image_axes=image_axes, axis_ranges=None, **kwargs)
+
+
+    def _sanitize_axis_ranges(self, *args):
+        """
+        This overrides the behaviour of ArrayAnimator to generate axis_ranges
+        based on the WCS.
+        """
+
+        def partial_pixel_to_world(pixel_dimension, pixel_coord):
+            wcs_dimension = self.wcs.pixel_n_dim - pixel_dimension - 1
+            corr = self.wcs.axis_correlation_matrix[:, wcs_dimension]
+            # If more than one world axis is linked to this dimension we can't
+            # display the world coordinate because we have no way of picking,
+            # so we just display pixel index.
+            if len(np.nonzero(corr)[0]) != 1:
+                return pixel_coord
+
+            slices = [0] * self.wcs.pixel_n_dim
+            slices[pixel_dimension] = slice(None)
+            # We know only one world axis correlates, so this should always
+            # return one world result.
+            return SlicedLowLevelWCS(self.wcs, slices).pixel_to_world_values(pixel_coord)
+
+
+        axis_ranges = [None] * self.wcs.pixel_n_dim
+        for i in self.slider_axes:
+            axis_ranges[i] = partial(partial_pixel_to_world, i)
+
+        return axis_ranges, None
+
+    def _get_main_axes(self):
+        axes = self.fig.add_axes([0.1, 0.1, 0.8, 0.8], projection=self.wcs,
+                                 slices=self.slices_wcsaxes)
+        return axes
+
+    def plot_start_image(self, ax):
[…]
