### 2019-10-14 4b1a218
pep8 fixes; add xrt and swap factory tests
