+++ sunpy/map/tests/test_map_factory.py
@@ -31,14 +31,13 @@ RHESSI_IMAGE = os.path.join(filepath, 'hsi_image_20101016_191218.fits')
 #==============================================================================
 class TestMap:
     def test_mapsequence(self):
-        #Test making a MapSequence
+        # Test making a MapSequence
         sequence = sunpy.map.Map(a_list_of_many, sequence=True)
         assert isinstance(sequence, sunpy.map.MapSequence)
 
     def test_composite(self):
-        #Test making a CompositeMap
-        comp = sunpy.map.Map(AIA_171_IMAGE, RHESSI_IMAGE,
-                         composite=True)
+        # Test making a CompositeMap
+        comp = sunpy.map.Map(AIA_171_IMAGE, RHESSI_IMAGE, composite=True)
         assert isinstance(comp, sunpy.map.CompositeMap)
 
     def test_patterns(self):
@@ -52,7 +51,7 @@ class TestMap:
         directory = os.path.join(filepath, "EIT")
         maps = sunpy.map.Map(directory)
         assert isinstance(maps, list)
-        assert ([isinstance(amap,sunpy.map.GenericMap) for amap in maps])
+        assert ([isinstance(amap, sunpy.map.GenericMap) for amap in maps])
         # Test that returned maps are sorted
         files_sorted = sorted([os.path.join(directory, f) for f in os.listdir(directory)])
         maps_sorted = [sunpy.map.Map(f) for f in files_sorted]
@@ -62,7 +61,7 @@ class TestMap:
         pattern = os.path.join(filepath, "EIT", "*")
         maps = sunpy.map.Map(pattern)
         assert isinstance(maps, list)
-        assert ([isinstance(amap,sunpy.map.GenericMap) for amap in maps])
+        assert ([isinstance(amap, sunpy.map.GenericMap) for amap in maps])
         # Test that returned maps are sorted
         files_sorted = sorted(glob.glob(pattern))
         maps_sorted = [sunpy.map.Map(f) for f in files_sorted]
@@ -75,7 +74,7 @@ class TestMap:
         # A list of filenames
         maps = sunpy.map.Map(a_list_of_many)
         assert isinstance(maps, list)
-        assert ([isinstance(amap,sunpy.map.GenericMap) for amap in maps])
+        assert ([isinstance(amap, sunpy.map.GenericMap) for amap in maps])
 
         # Data-header pair in a tuple
         pair_map = sunpy.map.Map((amap.data, amap.meta))
@@ -137,7 +136,7 @@ class TestMap:
         assert isinstance(amap, sunpy.map.GenericMap)
 
     def test_save(self):
-        #Test save out
+        # Test save out
         eitmap = sunpy.map.Map(a_fname)
         afilename = tempfile.NamedTemporaryFile(suffix='fits').name
         eitmap.save(afilename, filetype='fits', overwrite=True)
@@ -148,48 +147,54 @@ class TestMap:
 # Sources Tests
 #==============================================================================
     def test_sdo(self):
-        #Test an AIAMap
+        # Test an AIAMap
         aia = sunpy.map.Map(AIA_171_IMAGE)
-        assert isinstance(aia,sunpy.map.sources.AIAMap)
-        #Test a HMIMap
+        assert isinstance(aia, sunpy.map.sources.AIAMap)
+        # TODO: Test a HMIMap
 
     def test_soho(self):
[…]
