### 2019-10-14 d75bd1e
Better unit support in slider labels
