### 2019-10-14 bbdd62a
Cleanup and docstring
