### 2019-10-15 08ec35d
Rewrite the pixel to world function to not used SlicedLoweLevelWCS
This keeps compatibility with Astropy 3.1 for FITS WCS
