### 2019-10-15 234b419
Deprecate ImageAnimatorWCS
