### 2019-10-15 bdf4500
fix glob; less pathlib
