### 2019-10-15 0549ff0
Make tests start without matplotlib
