### 2019-10-15 4afe9ad
Move sunpy.cm to sunpy.visualization.colormaps
