### 2019-10-15 5a3c558
Move sunpy.cm to sunpy.visualization.colormaps
