### 2019-10-16 be29ee6
test for pathlib input
