### 2019-10-16 3d94f91
Give Map cmap and norm properties
