+++ sunpy/map/sources/sdo.py
@@ -55,8 +55,14 @@ class AIAMap(GenericMap):
         # Fill in some missing info
         self.meta['detector'] = self.meta.get('detector', "AIA")
         self._nickname = self.detector
-        self.plot_settings['cmap'] = plt.get_cmap(self._get_cmap_name())
-        self.plot_settings['norm'] = ImageNormalize(stretch=source_stretch(self.meta, AsinhStretch(0.01)))
+
+    @property
+    def cmap(self):
+        return plt.get_cmap(self._get_cmap_name())
+
+    @property
+    def norm(self):
+        return ImageNormalize(stretch=source_stretch(self.meta, AsinhStretch(0.01)))
 
     @property
     def _supported_observer_coordinates(self):
