+++ sunpy/map/sources/suvi.py
@@ -81,10 +81,15 @@ class SUVIMap(GenericMap):
         self.meta["detector"] = "SUVI"
         self.meta["telescop"] = "GOES-R"
         self._nickname = self.detector
-        self.plot_settings["cmap"] = plt.get_cmap(self._get_cmap_name())
-        self.plot_settings["norm"] = ImageNormalize(
-            stretch=source_stretch(self.meta, AsinhStretch(0.01))
-        )
+
+    @property
+    def cmap(self):
+        return plt.get_cmap(self._get_cmap_name())
+
+    @property
+    def norm(self):
+        return ImageNormalize(
+            stretch=source_stretch(self.meta, AsinhStretch(0.01)))
 
     @property
     def _supported_observer_coordinates(self):
