+++ sunpy/map/sources/hinode.py
@@ -64,7 +64,10 @@ class XRTMap(GenericMap):
         self.meta['detector'] = "XRT"
 #        self.meta['instrume'] = "XRT"
         self.meta['telescop'] = "Hinode"
-        self.plot_settings['cmap'] = plt.get_cmap(name='hinodexrt')
+
+    @property
+    def cmap(self):
+        return plt.get_cmap(name='hinodexrt')
 
     @property
     def measurement(self):
@@ -130,7 +133,9 @@ class SOTMap(GenericMap):
                  'SOT/SP': 'intensity',  # For the 1st 2 dimensions
                  }
 
-        self.plot_settings['cmap'] = plt.get_cmap('hinodesot' + color[self.instrument])
+    @property
+    def cmap(self):
+        return plt.get_cmap('hinodesot' + color[self.instrument])
 
     @classmethod
     def is_datasource_for(cls, data, header, **kwargs):
