+++ sunpy/map/sources/stereo.py
@@ -36,14 +36,20 @@ class EUVIMap(GenericMap):
 
         GenericMap.__init__(self, data, header, **kwargs)
         self._nickname = "{}-{}".format(self.detector, self.observatory[-1])
-        self.plot_settings['cmap'] = plt.get_cmap('sohoeit{wl:d}'.format(wl=int(self.wavelength.value)))
-        self.plot_settings['norm'] = ImageNormalize(stretch=source_stretch(self.meta, PowerStretch(0.25)))
         self.meta['waveunit'] = 'Angstrom'
 
         # Try to identify when the FITS meta data does not have the correct
         # date FITS keyword
         if ('date_obs' in self.meta) and not('date-obs' in self.meta):
             self.meta['date-obs'] = self.meta['date_obs']
+
+    @property
+    def cmap(self):
+        return plt.get_cmap('sohoeit{wl:d}'.format(wl=int(self.wavelength.value)))
+
+    @property
+    def norm(self):
+        return ImageNormalize(stretch=source_stretch(self.meta, PowerStretch(0.25)))
 
     @property
     def rsun_arcseconds(self):
