### 2019-10-16 1fe6908
Store colormap names instead of Colormap objects
