+++ sunpy/map/sources/rhessi.py
@@ -4,8 +4,6 @@
 __author__ = "Steven Christe"
 __email__ = "steven.d.christe@nasa.gov"
 
-import matplotlib.pyplot as plt
-
 from sunpy.map import GenericMap
 
 
@@ -57,7 +55,7 @@ class RHESSIMap(GenericMap):
 
         self.meta['waveunit'] = 'keV'
         self.meta['wavelnth'] = [self.meta['energy_l'], self.meta['energy_h']]
-        self.plot_settings['cmap'] = plt.get_cmap('rhessi')
+        self.plot_settings['cmap'] = 'rhessi'
 
     @property
     def detector(self):
