### 2019-10-16 f0aa895
Store colormap names instead of Colormap objects
