### 2019-10-16 c0a7e65
better glob check; comment on os.fspath
