### 2019-10-16 5295d7d
Significant improvements to the coordinates code reference
