+++ docs/code_ref/coordinates.rst
@@ -3,14 +3,90 @@
 SunPy coordinates
 =================
 
-The SunPy coordinates submodule is an implementation of the common solar physics
-coordinate frames using the :ref:`Astropy coordinates framework <astropy:astropy-coordinates>`.
+This sub-package contains:
+
+* A robust framework for working with solar-physics coordinate systems
+* Functions to obtain the locations of solar-system bodies (`sunpy.coordinates.ephemeris`)
+* Functions to calculate Sun-specific coordinate information (`sunpy.coordinates.sun`)
+
+The SunPy coordinate framework extends the
+:ref:`Astropy coordinates framework <astropy:astropy-coordinates>`.
+
+Supported Coordinate Systems
+----------------------------
+
+.. list-table::
+   :widths: auto
+   :header-rows: 1
+
+   * - Coordinate system
+     - Abbreviation
+     - SunPy/Astropy equivalent
+     - Notes
+   * - Heliocentric Aries Ecliptic (Mean)
+     - HAE (also HEC)
+     - Astropy's `~astropy.coordinates.HeliocentricMeanEcliptic`
+     -
+   * - Heliocentric Cartesian
+     - HCC
+     - `~sunpy.coordinates.frames.Heliocentric`
+     -
+   * - Heliocentric Earth Ecliptic
+     - HEE
+     - `~sunpy.coordinates.frames.HeliocentricEarthEcliptic`
+     -
+   * - Heliocentric Earth Equatorial
+     - HEEQ (also HEQ)
+     - `~sunpy.coordinates.frames.HeliographicStonyhurst`
+     - Use a Cartesian representation
+   * - Heliocentric Inertial
+     - HCI
+     - `~sunpy.coordinates.frames.HeliocentricInertial`
+     -
+   * - Heliocentric Radial
+     - HCR
+     - similar to `~sunpy.coordinates.frames.Heliocentric`
+     - Use a cylindrical representation, *but* with a 90-degree offset in ``psi``
+   * - Heliocentric/Heliographic Radial-Tangential-Normal
+     - HGRTN
+     - similar to `~sunpy.coordinates.frames.Heliocentric`
+     - The axes are permuted, with HCC X, Y, Z equivalent respectively to HGRTN Y, Z, X
+   * - Heliocentric Carrington
+     - HGC
+     - `~sunpy.coordinates.frames.HeliographicCarrington`
+     -
+   * - Heliocentric Stonyhurst
+     - HGS
+     - `~sunpy.coordinates.frames.HeliographicStonyhurst`
+     -
+   * - Helioprojective Cartesian
+     - HPC
+     - `~sunpy.coordinates.frames.Helioprojective`
+     -
+   * - Geocentric Earth Equatorial (Mean)
+     - GEI
+     - `~sunpy.coordinates.frames.GeocentricEarthEquatorial`
+     -
+   * - Geographic
+     - GEO
+     - Astropy's `~astropy.coordinates.ITRS`
+     - The precise geographic definitions may differ
+   * - Geocentric Solar Ecliptic
+     - GSE
+     - `~sunpy.coordinates.frames.GeocentricSolarEcliptic`
+     -
+
+
+For a description of these coordinate systems,
+see `Thompson (2006) <https://doi.org/10.1051/0004-6361:20054262>`_
+and `Franz & Harper (2002) <https://doi.org/10.1016/S0032-0633(01)00119-2>`_
+(and `corrected version <https://www2.mps.mpg.de/homes/fraenz/systems/systems3art/systems3art.html>`_).
 
 
 Getting Started
 ---------------
 
-The easiest interface to the coordinates module is through the `~astropy.coordinates.SkyCoord` class::
[…]
