### 2019-10-16 49e1c5e
tests for glob patterns besides *
