### 2019-10-16 6fa2bf7
Add changelogs/fix examples
