### 2019-10-16 13a9000
Add changelogs/fix examples
