### 2019-10-16 992d333
Correctly document TimeSeries peek() args/kwargs
