### 2019-10-16 0507c64
Correctly document TimeSeries peek() args/kwargs
