+++ sunpy/timeseries/sources/eve.py
@@ -61,6 +61,15 @@ class ESPTimeSeries(GenericTimeSeries):
 
     @peek_show
     def peek(self, title='EVE/ESP Level1', **kwargs):
+        """
+        Parameters
+        ----------
+        title : str, optional
+            Plot title.
+        **kwargs : dict
+            Additional plot keyword arguments that are handed to
+            :meth:`pandas.DataFrame.plot`.
+        """
 
         self._validate_data_for_plotting()
 
@@ -180,9 +189,10 @@ class EVESpWxTimeSeries(GenericTimeSeries):
         Parameters
         ----------
         column : `str`, optional
-            The column to display. Defaults to `None`, so it will display all.
+            The column to display. Defaults to ``None``, so it will display all.
         **kwargs : `dict`
-            Any additional plot arguments that should be used when plotting.
+            Additional plot keyword arguments that are handed to
+            :meth:`pandas.DataFrame.plot`.
         """
         # Check we have a timeseries valid for plotting
         self._validate_data_for_plotting()
