### 2019-10-17 e70bf26
Support pathlib.Path input to map_factory
