### 2019-10-17 e7a7fd7
Adding permission denied error in VSO
