### 2019-10-17 65d35fb
Adding permission denied error in VSO
