### 2019-10-17 2a6dfd5
Removing extra things
