### 2019-10-17 3c90a9f
Solving pep8 issue
