### 2019-10-17 c0970a1
Solving pep8 issue
