+++ changelog/3408.feature.2.rst
@@ -1,0 +1,2 @@
+Single character wildcards and character ranges can now be passed as
+glob patterns to `~sunpy.map.Map`.
