### 2019-10-17 bffe2cc
add logger; document pathlib usage
