### 2019-10-18 #1789 @mwcraig
@arfon -- I'd be delighted but fear I've got a conflict of interest here. Stuart and I are co-authors on the most recent astropy paper, but perhaps more importantly, have worked together on some of the pieces of Astropy that interface most directly with SunPy.
