### 2019-10-18 #1789 @mwcraig
I think I could be unbiased (the JOSS review criteria are pretty straightforward) but felt obligated to make you aware of al the appearance of a conflict of interest.
