### 2019-10-19 bd2b99f
Update changelog/3351.bugfix.rst
