### 2019-10-20 64afd3f
Adding error message in FIDO.fetch()
