### 2019-10-20 8791500
Adding error message in FIDO.fetch()
