### 2019-10-20 9bf7fdf
Adding error message in FIDO.fetch()
