### 2019-10-21 7c69dc0
Explicitly sort maps for glob and directory patterns (#3408)
Sort lists of maps from glob and directory; support more glob patterns in MapFactory; support passing pathlib.Path directly to MapFactory
