+++ changelog/3371.bugfix.rst
@@ -1,1 +1,0 @@
-Fixing the ordering of lon and lat inputs into `sunpy.map.make_fitswcs_header`
