### 2019-10-22 014cc3d
Reduce the number of files we upload after doc build
This should hopefully speed up the build
