+++ tox.ini
@@ -41,7 +41,7 @@ deps =
     astropy
     astroquery
     jplephem
-commands = sphinx-build docs docs/_build/html -W -b html
+commands = sphinx-build docs docs/_build/html -W -b html -d docs/_build/.doctrees
 
 # This env requires tox-conda.
 [testenv:figure]
