### 2019-10-22 d6571c3
Fix the manifest file to include the right helpers things and docs
