### 2019-10-22 591bde0
update stability and add colormaps plot
