+++ sunpy/visualization/colormaps/__init__.py
@@ -1,5 +1,13 @@
 """
-The following colormaps are provided by this module:
+The following colormaps are provided by this module.
+
+.. plot::
+
+    import matplotlib.pyplot as plt
+    import sunpy.visualization.colormaps as cm
+    cm.show_colormaps()
+
+
 """
 from sunpy.visualization.colormaps.cm import *  # noqa
 
