### 2019-10-22 5029b68
update stability and add colormaps plot
