### 2019-10-23 4048bea
Adding test for the Error
