### 2019-10-23 35060cb
Adding test for the Error
