+++ changelog/3410.removal.rst
@@ -1,0 +1,2 @@
+``sunpy.cm`` has been moved to `sunpy.visualization.colormaps` and will be
+removed in a future version.
