### 2019-10-23 4568cea
Remove redundant doctest line
Co-Authored-By: Albert Y. Shih <ayshih@gmail.com>
