### 2019-10-23 2758851
Removing some styling mistakes
