+++ changelog/3417.bugfix .rst
@@ -1,1 +1,1 @@
-Fixed a bug permission denied errors on VSO results are ugly  by adding an error message in FIDO.fetch().
+Fixed a bug where permission denied errors when downloading files are very verbose by adding an error message in `~sunpy.net.fido_factory.UnifiedDownloaderFactory.fetch`.
