### 2019-10-23 d051bcb
Removing some styling mistakes
