### 2019-10-23 #1789 @arfon
> I think I could be unbiased (the JOSS review criteria are pretty straightforward) but felt obligated to make you aware of al the appearance of a conflict of interest.

@mwcraig - Thanks again for the offer to help here. I'd like to take you up on your offer :-)
