### 2019-10-23 #1832 @Cadair
The author list is currently not in the right format. I am still soliciting ORCIDs and Affiliations from people, I will reformat the author list and create a zenodo author file at some point soon. I can do this at whatever point you feel is most appropriate, but am stalling to give people time to give me their info.
