### 2019-10-23 8eadf16
Apply suggestions from code review
Co-Authored-By: Punyaslok Pattnaik <punyaslokpattnaik@yahoo.co.in>
Co-Authored-By: Benjamin Mampaey <benjamin.mampaey@oma.be>
