### 2019-10-23 #1832 @Cadair
I fixed the ORCID, but that error is because the author list is just in the wrong format.
