### 2019-10-23 #1832 @arfon
@mwcraig, @ceb8 - for now, please use this copy of the paper for your review for now: https://github.com/openjournals/joss-papers/blob/joss.01789/joss.01789/10.21105.joss.01789.pdf
