### 2019-10-23 70d20e5
Explicitly set the clipping to False for the default normalizers for Map sources
