### 2019-10-23 922606a
Explicitly set the clipping to False for the default normalizers for Map sources
