+++ sunpy/map/sources/trace.py
@@ -61,7 +61,7 @@ class TRACEMap(GenericMap):
         self._nickname = self.detector
         # Colour maps
         self.plot_settings['cmap'] = plt.get_cmap('trace' + str(self.meta['WAVE_LEN']))
-        self.plot_settings['norm'] = ImageNormalize(stretch=source_stretch(self.meta, LogStretch()))
+        self.plot_settings['norm'] = ImageNormalize(stretch=source_stretch(self.meta, LogStretch()), clip=False)
 
     @classmethod
     def is_datasource_for(cls, data, header, **kwargs):
