+++ sunpy/map/sources/mlso.py
@@ -44,7 +44,7 @@ class KCorMap(GenericMap):
         self._nickname = self.detector
 
         self.plot_settings['cmap'] = plt.get_cmap(self._get_cmap_name())
-        self.plot_settings['norm'] = ImageNormalize(stretch=source_stretch(self.meta, PowerStretch(0.25)))
+        self.plot_settings['norm'] = ImageNormalize(stretch=source_stretch(self.meta, PowerStretch(0.25)), clip=False)
         # Negative value pixels can appear that lead to ugly looking images.
         # This can be fixed by setting the lower limit of the normalization.
         self.plot_settings['norm'].vmin = 0.0
