### 2019-10-23 5c8d558
Explicitly set the clipping to False for the default normalizers for Map sources
