### 2019-10-23 5af776e
Correct deprecation and doc
