### 2019-10-23 2fcad16
Move wcs to it's own file
