### 2019-10-23 b1bde65
Move wcs to it's own file
