### 2019-10-23 9579c10
fixing timerange issue for scraper
