### 2019-10-23 5e8e630
Add missing language corrections
