### 2019-10-23 9af9e33
Add missing language corrections
