### 2019-10-23 810db25
Add missing language corrections
