### 2019-10-23 93908dc
Make conversion errors more helpful
