### 2019-10-23 bfee992
changing spelling of plotting
