+++ sunpy/timeseries/timeseriesbase.py
@@ -410,7 +410,7 @@ class GenericTimeSeries:
             Any additional plot arguments that should be used when plotting.
         """
         # Check we have a timeseries valid for plotting
-        self._validate_data_for_ploting()
+        self._validate_data_for_plotting()
 
         # Now make the plot
         figure = plt.figure()
@@ -418,7 +418,7 @@ class GenericTimeSeries:
 
         return figure
 
-    def _validate_data_for_ploting(self):
+    def _validate_data_for_plotting(self):
         """
         Raises an exception if the `~sunpy.timeseries.TimeSeries` is invalid
         for plotting.
