### 2019-10-24 aff0f7d
Update sunpy/cm/__init__.py
Co-Authored-By: Jack Ireland <jackireland@gmail.com>
