### 2019-10-24 8a50b33
Do a nice little hack to allow imports from old modules
