### 2019-10-24 454930f
Added changelog and tests for clipping change
