### 2019-10-24 04224e0
Added changelog and tests for clipping change
