+++ changelog/3427.bugfix.rst
@@ -1,0 +1,1 @@
+Fixed a bug where clipping behavior had been enabled by default in the plotting normalizers for ``Map`` objects.  Clipping needs to be disabled to make use of the over/under/masked colors in the colormap.
