### 2019-10-24 ecd4d87
Added changelog and tests for clipping change
