### 2019-10-24 4c93bdf
Put parse_time example in an example folder
