### 2019-10-24 12b1f8b
Change log file for PR #3429
