+++ changelog/3429.trivial.rst
@@ -1,0 +1,1 @@
+Corrected spelling of 'plotting' in timeseries method (changed 'ploting' to 'plotting'). 
