+++ sunpy/database/tests/test_tables.py
@@ -529,21 +529,21 @@ def test_entry_from_query_results_with_none_wave_and_default_unit(
             source='SOHO', provider='SDAC', physobs='intensity',
             fileid='/archive/soho/private/data/processed/virgo/spm/VIRGO-SPM-BLUE-L2-MISSIONLONG.fits',
             observation_time_start=datetime(1996, 1, 23, 0, 0),
-            observation_time_end=datetime(2019, 5, 5, 23, 59),
+            observation_time_end=datetime(2019, 9, 30, 23, 59),
             instrument='VIRGO', size=32652.0, wavemin=None,
             wavemax=None),
         DatabaseEntry(
             source='SOHO', provider='SDAC', physobs='intensity',
             fileid='/archive/soho/private/data/processed/virgo/spm/VIRGO-SPM-GREEN-L2-MISSIONLONG.fits',
             observation_time_start=datetime(1996, 1, 23, 0, 0),
-            observation_time_end=datetime(2019, 5, 5, 23, 59),
+            observation_time_end=datetime(2019, 9, 30, 23, 59),
             instrument='VIRGO', size=32652.0, wavemin=None,
             wavemax=None),
         DatabaseEntry(
             source='SOHO', provider='SDAC', physobs='intensity',
             fileid='/archive/soho/private/data/processed/virgo/spm/VIRGO-SPM-RED-L2-MISSIONLONG.fits',
             observation_time_start=datetime(1996, 1, 23, 0, 0),
-            observation_time_end=datetime(2019, 5, 5, 23, 59),
+            observation_time_end=datetime(2019, 9, 30, 23, 59),
             instrument='VIRGO', size=32652.0, wavemin=None,
             wavemax=None),
         DatabaseEntry(
