+++ setup.cfg
@@ -91,7 +91,7 @@ all_files = 1
 minversion = 3.0
 testpaths = "sunpy" "docs"
 # Skip sunpy/data to prevent importing the sample data (there are no tests in that dir anyway)
-norecursedirs = ".tox" "build" "docs[\/]_build" "docs[\/]generated" "*.egg-info" "astropy_helpers" "examples" "sunpy[\/]data"
+norecursedirs = ".tox" "build" "docs[\/]_build" "docs[\/]generated" "*.egg-info" "astropy_helpers" "examples" "sunpy[\/]data" ".jupyter"
 doctest_plus = enabled
 doctest_optionflags = NORMALIZE_WHITESPACE FLOAT_CMP ELLIPSIS
 addopts = -p no:warnings --doctest-rst -m "not figure"
