+++ azure-pipelines.yml
@@ -18,6 +18,8 @@ trigger:
   branches:
     include:
     - '*'
+    exclude:
+    - '*backport*'
   tags:
     include:
     - 'v*'
