### 2019-10-25 055f1ae
Change CircleCI logic
