### 2019-10-25 e0f6de6
Change CircleCI logic
