### 2019-10-25 e9c2dd8
Updated the guide to use clip_interval
