### 2019-10-25 4a2ca0f
Updated the guide to use clip_interval
