+++ changelog/3450.doc.rst
@@ -1,0 +1,1 @@
+Updated the user guide for Map to use ``clip_interval``.
