+++ changelog/3451.doc.rst
@@ -1,0 +1,1 @@
+Updated the Venus-transit gallery to use the VSO so that it has correct pointing information in the header.
