+++ setup.cfg
@@ -29,7 +29,9 @@ classifiers =
 python_requires = >=3.6
 packages = find:
 include_package_data = True
-setup_requires = setuptools_scm
+setup_requires =
+  setuptools_scm
+  importlib_resources;python_version<"3.7"
 install_requires =
   numpy>=1.14.5
   scipy
@@ -37,6 +39,7 @@ install_requires =
   pandas
   astropy>=3.1
   parfive[ftp]
+  importlib_resources;python_version<"3.7"
 
 [options.extras_require]
 database = sqlalchemy
