+++ sunpy/coordinates/tests/test_utils.py
@@ -6,17 +6,22 @@ from astropy.coordinates import SkyCoord
 import astropy.units as u
 
 from sunpy.coordinates import sun
-import sunpy.map
-import sunpy.data.test
+import sunpy.map as smap
+import sunpy.data.test as test
 from sunpy.coordinates import frames
 from sunpy.coordinates.utils import GreatArc
 
+
+@pytest.fixture
+def aia171_test_map():
+    return smap.Map(test.get_test_filepath('aia_171_level1.fits'))
 
 # Test the great arc code against calculable quantities
 # The inner angle is the same between each pair of co-ordinates.  You can
 # calculate these co-ordinates using the inner angle formulae as listed here:
 # https://en.wikipedia.org/wiki/Great-circle_distance
-#
+
+
 @pytest.mark.parametrize("start, end", [((0, 0), (0, 45)),
                                         ((0, 0), (45, 0)),
                                         ((0, 45), (0, 0)),
@@ -32,23 +37,26 @@ def test_great_arc_calculable(start, end):
                  observer=frames.HeliographicStonyhurst(0*u.deg, 0*u.deg, 1*u.AU))
     gc = GreatArc(c, d)
 
-    assert gc.start.x == c.transform_to(frames.Heliocentric).x
-    assert gc.start.y == c.transform_to(frames.Heliocentric).y
-    assert gc.start.z == c.transform_to(frames.Heliocentric).z
+    c_trans = c.transform_to(frames.Heliocentric)
+    assert gc.start.x == c_trans.x
+    assert gc.start.y == c_trans.y
+    assert gc.start.z == c_trans.z
     assert gc.start.observer.lat == 0*u.deg
     assert gc.start.observer.lon == 0*u.deg
     assert gc.start.observer.radius == 1 * u.AU
 
-    assert gc.end.x == d.transform_to(frames.Heliocentric(observer=c.observer)).x
-    assert gc.end.y == d.transform_to(frames.Heliocentric(observer=c.observer)).y
-    assert gc.end.z == d.transform_to(frames.Heliocentric(observer=c.observer)).z
+    d_trans = d.transform_to(frames.Heliocentric(observer=c.observer))
+    assert gc.end.x == d_trans.x
+    assert gc.end.y == d_trans.y
+    assert gc.end.z == d_trans.z
     assert gc.end.observer.lat == 0*u.deg
     assert gc.end.observer.lon == 0*u.deg
     assert gc.end.observer.radius == 1 * u.AU
 
     np.testing.assert_almost_equal(gc.inner_angle.to('deg').value, 45.0)
     np.testing.assert_almost_equal(gc.radius.to('km').value, sun.constants.radius.to('km').value)
-    np.testing.assert_almost_equal(gc.distance.to('km').value, sun.constants.radius.to('km').value * 2 * np.pi/8, decimal=1)
+    np.testing.assert_almost_equal(gc.distance.to(
+        'km').value, sun.constants.radius.to('km').value * 2 * np.pi/8, decimal=1)
 
 
 # Test the calculation of coordinates using varying numbers of points on
@@ -60,20 +68,22 @@ def test_great_arc_calculable(start, end):
                           (3, 3, (600, -600), (-100, 800), 1.8683580432741789, 1300377.1981299),
                           # Test equally spaced monotonically increasing numpy
                           # array
-                          (np.linspace(0, 1, 43), 43, (600, -600), (-100, 800), 1.8683580432741789, 1300377.1981299),
+                          (np.linspace(0, 1, 43), 43, (600, -600),
+                           (-100, 800), 1.8683580432741789, 1300377.1981299),
[…]
