### 2019-10-27 4dc8a36
Update setup.cfg
