+++ setup.cfg
@@ -31,7 +31,6 @@ packages = find:
 include_package_data = True
 setup_requires =
   setuptools_scm
-  importlib_resources;python_version<"3.7"
 install_requires =
   numpy>=1.14.5
   scipy
