### 2019-10-28 3eb4bba
Update setup.cfg
