+++ setup.cfg
@@ -38,7 +38,7 @@ install_requires =
   pandas
   astropy>=3.1
   parfive[ftp]
-  importlib_resources;python_version<"3.7"
+  importlib_resources;python_version<"3.8"
 
 [options.extras_require]
 database = sqlalchemy
