### 2019-10-28 e73773c
Fixes to Map plotting in user guide
