### 2019-10-28 a8030bc
Fixes to Map plotting in user guide
