+++ changelog/3454.doc.rst
@@ -1,0 +1,1 @@
+Fixed the plots with multiple subplots in the ``Map`` user guide to properly use `~astropy.visualization.wcsaxes` and to be appropriately sized.
