### 2019-10-28 c46c1cb
Fix transforming an array of obstimes to a single obstime
