### 2019-10-28 31ce606
Fix a deprecation warning in rotate
