### 2019-10-28 8dde7e9
ignore sunpy.cm when running tests
