### 2019-10-28 34c36e7
Update changelog/3378.breaking.rst
Co-Authored-By: Stuart Mumford <stuart@cadair.com>
