+++ changelog/3378.breaking.rst
@@ -1,2 +1,2 @@
-:meth:`NOAAIndicesTimeSeries.peek()` now checks that the `type` argument is a
+`NOAAIndicesTimeSeries.peek` now checks that the `type` argument is a
 valid string, and raises a `ValueError` if it isn't.
