### 2019-10-28 047fac3
move partial_pixel_to_world to its own method
