### 2019-10-28 3f3509e
Add a non-plot test for ArrayAnimatorWCS
