### 2019-10-28 5f974b4
Add a non-plot test for ArrayAnimatorWCS
