+++ changelog/3407.feature.rst
@@ -1,0 +1,5 @@
+`sunpy.visualization.animator.ArrayAnimatorWCS` was added which uses the WCS
+object for all operations, including the slider labels. It also provides the
+ability to customise the plot by specifying arguments to
+`~astropy.visualization.wcsaxes.WCSAxes` methods and supports animation of
+WCS aware line plots with Astroy 4.0.
