### 2019-10-29 43ac9fe
Write a functioning read only test
