### 2019-10-29 ab01562
Undo circle changes
