+++ .circleci/config.yml
@@ -45,7 +45,6 @@ jobs:
     steps:
       - checkout
       - run: *skip-check
-      - run: *no-backports
       - run: *merge-check
       - run: python setup.py egg_info
 
@@ -55,7 +54,6 @@ jobs:
     steps:
       - checkout
       - run: *skip-check
-      - run: *no-backports
       - run: *merge-check
       - run: python setup.py egg_info
 
@@ -65,7 +63,6 @@ jobs:
     steps:
       - checkout
       - run: *skip-check
-      - run: *no-backports
       - run: *merge-check
       - run: *permission-run
       - run: python setup.py sdist
@@ -78,7 +75,6 @@ jobs:
     steps:
       - checkout
       - run: *skip-check
-      - run: *no-backports
       - run: *merge-check
       - run: *permission-run
       - run: *sudo-apt-install
@@ -98,7 +94,6 @@ jobs:
           keys: sample-data-v1
       - checkout
       - run: *skip-check
-      - run: *no-backports
       - run:
           name: apt dependencies
           command: |
@@ -114,7 +109,6 @@ jobs:
     steps:
       - checkout
       - run: *skip-check
-      - run: *no-backports
       - run: *merge-check
       - run: *apt-install
       - run: pip install -U tox tox-conda codecov
@@ -134,7 +128,6 @@ jobs:
           keys: sample-data-v1
       - checkout
       - run: *skip-check
-      - run: *no-backports
       - run: *merge-check
       - run: *permission-run
       - run: *sudo-apt-install
