### 2019-10-29 64f11fc
Add a thrid party examples section
