+++ examples/sunpy_other_packages/README.txt
@@ -1,0 +1,4 @@
+Using SunPy with other packages
+===============================
+
+These examples show how SunPy can interface with other packages to perform various tasks.
