### 2019-10-29 e63ad75
Fix raising a warning about missing observer when not needed
