### 2019-10-29 ad2a3f9
Improve documentation for astroquery examples
