### 2019-10-29 a70a72b
Improve documentation for astroquery examples
