+++ examples/sunpy_other_packages/getting_lasco_observer_location.py
@@ -11,7 +11,7 @@ This requires the intallation of the
 `astroquery <https://astroquery.readthedocs.io/en/latest/>`__
 package and an internet connection.
 `astroquery <https://astroquery.readthedocs.io/en/latest/>`__ can be installed ontop of
-the existing sunpy conda environemnt: conda install -c astropy astroquery
+the existing sunpy conda environment: ``conda install -c astropy astroquery``
 """
 
 # sphinx_gallery_thumbnail_number = 2
