+++ examples/sunpy_other_packages/reprojection_heliographic_stonyhurst.py
@@ -1,10 +1,15 @@
 """
-==================
-Re-projecting Maps
-==================
+===========================
+Creating a Heliographic Map
+===========================
 
-How to reproject a map into
+In this example we use the :ref:`reproject <reproject:quickstart>` package to
+make a heliographic image from an AIA image.
+
+You will need to install reproject to run this example.
 """
+# sphinx_gallery_thumbnail_number = 2
+
 import matplotlib.pyplot as plt
 import numpy as np
 
@@ -14,8 +19,8 @@ from astropy.wcs import WCS
 
 from reproject import reproject_interp
 
-import sunpy.map
 import sunpy.data.sample
+import sunpy.map
 
 ###############################################################################
 # We start with the sample data
@@ -26,9 +31,14 @@ ax = plt.subplot(projection=aia_map)
 aia_map.plot(ax)
 
 ###############################################################################
-# Now construct a header for the output
+# Reproject works by transforming an input image (with a `~astropy.wcs.WCS`) to
+# a output image, specified by a different WCS object. Therefore we need to
+# build a `~astropy.wcs.WCS` object describing the output we desire.
+# To do this we use the `sunpy.map.make_fitswcs_header` which assists us in
+# constructing this WCS object. Here we create a WCS based on a heliographic
+# Stonyhurst reference coordinate and with the CAR (plate carrée) projection.
 shape_out = [720, 1440]
-header = sunpy.map.make_fitswcs_header(np.empty(shape_out),
+header = sunpy.map.make_fitswcs_header(shape_out,
                                        SkyCoord(0, 0, unit=u.deg,
                                                 frame="heliographic_stonyhurst",
                                                 obstime=aia_map.date),
@@ -40,8 +50,10 @@ out_wcs = WCS(header)
 
 ###############################################################################
 # With the new header, re-project the data into the new coordinate system.
-array, footprint = reproject_interp(
-    (aia_map.data, aia_map.wcs), out_wcs, shape_out=shape_out)
+# Here we are using the fastest but least accurate method of reprojection,
+# `reproject.reproject_interp`, a more accurate but slower method is
+# `reproject.reproject_adaptive`.
+array, footprint = reproject_interp(aia_map, out_wcs, shape_out=shape_out)
 outmap = sunpy.map.Map((array, header))
 outmap.plot_settings = aia_map.plot_settings
 
