### 2019-10-29 9ec1683
Adjust docs and astropydev dependencies
