+++ tox.ini
@@ -23,8 +23,9 @@ deps =
     online: pytest-rerunfailures
     online: pytest-timeout
     online: astroquery
-    astropydev: git+https://github.com/astropy/astroquery
-    astropydev: jplephem
+    astropydev,build_docs: reproject
+    astropydev,build_docs: astroquery
+    astropydev,build_docs: jplephem
 commands =
     offline,astropydev,numpydev,astropy31: {env:PYTEST_COMMAND} {posargs}
     online: {env:PYTEST_COMMAND} --reruns 2 --timeout=180 --remote-data=any {posargs}
@@ -36,12 +37,6 @@ changedir = {toxinidir}
 extras = dev
 setenv =
     HOME = {envtmpdir}
-deps =
-    # For now.
-    astropy
-    astroquery
-    jplephem
-    reproject
 commands = sphinx-build docs docs/_build/html -W -b html -d docs/_build/.doctrees
 
 # This env requires tox-conda.
