### 2019-10-30 2c4faf2
fixing tests with remote_data
