### 2019-10-30 f189576
fixing tests with remote_data
