### 2019-10-30 c3bb1dc
Add mosiac example
