+++ sunpy/cm/__init__.py
@@ -5,7 +5,7 @@ import warnings as _w
 from sunpy.util.exceptions import SunpyDeprecationWarning as _SDW
 
 _w.warn("The functionality of the sunpy.cm module is now in sunpy.visualization.colormaps as of SunPy 1.1. "
-        "The ability to import sunpy.cm as a module may be removed in a future version of SunPy.",
+        "The ability to import sunpy.cm as a module will be removed in a future version of SunPy.",
         category=_SDW)
 
 # Import docstring from the new module
