### 2019-10-30 ffea3c1
Apply suggestions from code review
Co-Authored-By: Albert Y. Shih <ayshih@gmail.com>
