### 2019-10-30 d7b9f95
Apply suggestions from code review
Co-Authored-By: Albert Y. Shih <ayshih@gmail.com>
