### 2019-10-30 f1fbcf4
Apply suggestions from code review
Co-Authored-By: Albert Y. Shih <ayshih@gmail.com>
