### 2019-10-31 7d9e338
Clean up map.save() docstring
