### 2019-10-31 4f531dd
Check that a Path works with sunpy.io.write_file
