+++ changelog/3469.feature.rst
@@ -1,0 +1,1 @@
+`sunpy.io.write_file()` now accepts `~pathlib.Path` objects as filename inputs.
