### 2019-10-31 a66c7cf
Remove u.Quantity calls from examples
