### 2019-10-31 b8d73dc
Remove u.Quantity calls from examples
