### 2019-11-01 2663a38
Remove u.Quantity calls from examples (#3470)
