### 2019-11-01 2baaf50
Bump test coverage up for colormaps
