### 2019-11-01 cb2ce20
fixing precision problem in scraper
