### 2019-11-01 184877c
Move the animator play button to the left side
This is so that the label can't disappear behind the button
