### 2019-11-01 77c69bb
apply expanduser before passing to glob in map factory
