### 2019-11-02 d60ca09
Implement changes from review.
