### 2019-11-02 fd17266
Implement changes from review.
