### 2019-11-03 cd79196
More authors
Co-Authored-By: Jack Ireland <jackireland@gmail.com>
Co-Authored-By: Pritish Chakraborty <chakrabortypritish@gmail.com>
Co-Authored-By: Yash Jain <yashjainjain1704@gmail.com>
Co-Authored-By: SallyDa <sallydacie@gmail.com>
Co-Authored-By: Serge Zahniy <ddialer@gmail.com>
Co-Authored-By: Nick Murphy <namurphy@cfa.harvard.edu>
Co-Authored-By: Kalpesh Krishna <kalpeshk2011@gmail.com>
Co-Authored-By: Mickaël Schoentgen <contact@tiger-222.fr>
Co-Authored-By: gyengen <n.g.gyenge@sheffield.ac.uk>
Co-Authored-By: Yudhik Agrawal <yudhik100@gmail.com>
Co-Authored-By: Sophie Murray <sophiemur
[…]
