### 2019-11-03 69e4611
Add __citation__
Based off https://www.astropy.org/acknowledging.html, this adds
a `__citation__` variable to the top-level of the package.
