+++ sunpy/__init__.py
@@ -37,6 +37,32 @@ try:
 except ImportError:
     __version__ = ''
 
+__citation__ = """\
+@ARTICLE{2015CS&D....8a4009S,
+   author = {{SunPy Community}, T. and {Mumford}, S.~J. and {Christe}, S. and
+    {P{\'e}rez-Su{\'a}rez}, D. and {Ireland}, J. and {Shih}, A.~Y. and
+    {Inglis}, A.~R. and {Liedtke}, S. and {Hewett}, R.~J. and {Mayer}, F. and
+    {Hughitt}, K. and {Freij}, N. and {Meszaros}, T. and {Bennett}, S.~M. and
+    {Malocha}, M. and {Evans}, J. and {Agrawal}, A. and {Leonard}, A.~J. and
+    {Robitaille}, T.~P. and {Mampaey}, B. and {Iv{\'a}n Campos-Rozo}, J. and
+    {Kirk}, M.~S.},
+    title = "{SunPy{\mdash}Python for solar physics}",
+  journal = {Computational Science and Discovery},
+archivePrefix = "arXiv",
+   eprint = {1505.02563},
+ primaryClass = "astro-ph.IM",
+     year = 2015,
+    month = jan,
+   volume = 8,
+   number = 1,
+      eid = {014009},
+    pages = {014009},
+      doi = {10.1088/1749-4699/8/1/014009},
+   adsurl = {http://adsabs.harvard.edu/abs/2015CS%26D....8a4009S},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}"""
+
+
 if not _SUNPY_SETUP_:
     from sunpy.util.config import load_config, print_config
     from sunpy.util import system_info
