### 2019-11-04 eb6c2ba
Add an astropydev figure test build
