+++ tox.ini
@@ -16,7 +16,7 @@ extras = all,tests
 deps =
     astropydev,numpydev: cython
     numpydev: git+https://github.com/numpy/numpy
-    astropydev: git+https://github.com/astropy/astropy
+    astropydev,figure_astropydev: git+https://github.com/astropy/astropy
     pytest-cov
     pytest-xdist
     astropy31: astropy<3.2
@@ -56,6 +56,22 @@ conda_deps =
     matplotlib = 3.1.1
     scipy = 1.2.1
     pandas = 0.24.2
+    pytest-cov
+conda_channels = conda-forge
+commands = {env:PYTEST_COMMAND} -m "figure" {posargs}
+
+# This env requires tox-conda.
+[testenv:figure_astropydev]
+changedir = {toxinidir}
+basepython = python3.7
+extras = all,tests
+conda_deps =
+    python = 3.7.4
+    numpy = 1.17.2
+    freetype = 2.10
+    matplotlib = 3.1.1
+    scipy = 1.3.1
+    pandas = 0.25.1
     pytest-cov
 conda_channels = conda-forge
 commands = {env:PYTEST_COMMAND} -m "figure" {posargs}
