### 2019-11-04 3051dec
Fix many WCS logging stuff
