### 2019-11-04 dcdab51
update grid for colorbar, and add buttons test
