+++ sunpy/visualization/animator/tests/test_wcs.py
@@ -113,6 +113,16 @@ def test_array_animator_wcs_2d_transpose_update_plot(wcs_4d):
 
 
 @figure_test
+def test_array_animator_wcs_2d_colorbar_buttons(wcs_4d):
+    data = np.arange(120).reshape((5, 4, 3, 2))
+    bf = [lambda x: x]*10
+    bl = ['h']*10
+    a = ArrayAnimatorWCS(data, wcs_4d, [0, 0, 'y', 'x'], colorbar=True, button_func=bf, button_labels=bl)
+    a.update_plot(1, a.im, a.sliders[0]._slider)
+    return a.fig
+
+
+@figure_test
 def test_array_animator_wcs_1d_update_plot(wcs_4d):
     pytest.importorskip("astropy", minversion="4.0dev26173")
     data = np.arange(120).reshape((5, 4, 3, 2))
