### 2019-11-04 15cd874
Test the output of the slider range callable
