### 2019-11-04 d2b0e64
Update docstring
