### 2019-11-04 e2534c6
Bump test coverage up for colormaps (#3474)
* Bump test coverage up for colormaps

* Update docstring
