### 2019-11-04 3ffc2a2
Bump test coverage up for colormaps (#3474)
* Bump test coverage up for colormaps

* Update docstring
