### 2019-11-05 0314cad
fixing precision issue with astropy.time and convert to datetime
