+++ sunpy/util/scraper.py
@@ -232,8 +232,8 @@ class Scraper:
                             fullpath = directory + href
                             if self._URL_followsPattern(fullpath):
                                 datehref = self._extractDateURL(fullpath)
-                                if (datehref - timerange.start) >= 0 and \
-                                   (datehref - timerange.end) <= 0:
+                                if (datehref.to_datetime() >= timerange.start.to_datetime() and
+                                    datehref.to_datetime() <= timerange.end.to_datetime()):
                                         filesurls.append(fullpath)
                 finally:
                     opn.close()
