### 2019-11-05 92ecfab
have to add satellite number as there is no such file for test time of interest for result of _get_goes_satellite
