### 2019-11-05 aa78287
fixing tests - old dates do not work with scraper as they do not exist!
