### 2019-11-05 e8073a8
Add an example of reprojecting between different observers
