+++ examples/sunpy_other_packages/reprojection_differet_observers.py
@@ -1,0 +1,190 @@
+"""
+Reprojecting AIA Images to Different Observers
+==============================================
+
+This example demonstrates how you can reproject images to the view from
+different observers, we use both AIA and STEREO A data to demonstrate this.
+
+"""
+# sphinx_gallery_thumbnail_number = 2
+
+import matplotlib.pyplot as plt
+import astropy.units as u
+from astropy.wcs import WCS
+from astropy.coordinates import SkyCoord
+
+import sunpy.map
+from sunpy.coordinates import get_body_heliographic_stonyhurst
+from sunpy.net import Fido, attrs as a
+
+from reproject import reproject_interp
+
+
+######################################################################
+# In this example we are going to make a lot of side by side figures, so
+# let’s change the default figure size.
+#
+
+plt.rcParams['figure.figsize'] = (16, 8)
+
+
+######################################################################
+# Let’s download an AIA and STEREO A where the separation is around 90
+# deg.
+#
+
+stereo = (a.vso.Source('STEREO_A') &
+          a.Instrument('EUVI') &
+          a.Time('2010-08-19', '2010-08-19T00:10:00'))
+
+aia = (a.Instrument('AIA') &
+       a.vso.Sample(24 * u.hour) &
+       a.Time('2010-08-19', '2010-08-19T00:10:00'))
+
+wave = a.Wavelength(17 * u.nm, 18 * u.nm)
+
+
+res = Fido.search(wave, aia | stereo)
+
+files = Fido.fetch(res)
+
+
+######################################################################
+# Create a map for each image.
+#
+
+map_aia, map_stereo = sunpy.map.Map(sorted(files))
+
+fig = plt.figure()
+ax1 = fig.add_subplot(1,2,1, projection=map_aia)
+map_aia.plot(axes = ax1)
+ax2 = fig.add_subplot(1,2,2, projection=map_stereo)
+map_stereo.plot(axes = ax2)
+plt.show()
+
+######################################################################
+# We now need to construct an output WCS. Because we want to downsample
+# the image (for performance) we build a custom header using
+# ``sunpy.map.make_fitswcs_header`` but we use a lot of the ``map_aia``
+# properties to do it. We use the reference coordinate from the AIA image,
+# and make the scale 4x larger to compensate for the fact we are making
+# the resolution 4x lower.
+#
+
+out_shape = (1024, 1024)
+out_header = sunpy.map.make_fitswcs_header(
+    out_shape,
+    map_aia.reference_coordinate,
+    scale=u.Quantity(map_aia.scale)*4,
+    instrument="EUVI",
+    observatory="AIA Observer",
+    wavelength=map_stereo.wavelength
+)
+
+
+######################################################################
+# Next we construct an ``~astropy.wcs.WCS`` object from the header.
+# Currently ``~astropy.wcs.WCS`` does not understand the observer
+# position, so we manually set that.
+#
+
+out_wcs = WCS(out_header)
+out_wcs.heliographic_observer = map_aia.reference_coordinate.observer
+
+
+######################################################################
+# We can now reproject the stereo map to this output ``~astropy.wcs.WCS``.
+#
+# Here we are using the fastest but least accurate method of reprojection,
+# `reproject.reproject_interp`, a more accurate but slower method is
+# `reproject.reproject_adaptive`.
+
+output, footprint = reproject_interp(map_stereo, out_wcs, out_shape)
+
+
+######################################################################
+# We can now plot the STEREO image as seen from the position of SDO, next
+# to the AIA image.
+#
+
+outmap = sunpy.map.Map(output, out_header)
+outmap.plot_settings = map_stereo.plot_settings
+
+fig = plt.figure()
[…]
