### 2019-11-05 66861e4
final fixes of tests to take into account scraping web
