+++ sunpy/net/dataretriever/tests/test_lyra_ud.py
@@ -10,7 +10,7 @@ from sunpy.net import Fido
 from sunpy.net import attrs as a
 
 from hypothesis import given, settings
-from sunpy.net.tests.strategies import time_attr
+from sunpy.net.tests.strategies import range_time
 
 LCClient = lyra.LYRAClient()
 
@@ -37,7 +37,7 @@ def test_get_url_for_time_range(timerange, url_start, url_end):
     assert urls[-1] == url_end
 
 
-@given(time_attr())
+@given(range_time('2010-01-06'))
 def test_can_handle_query(time):
     ans1 = lyra.LYRAClient._can_handle_query(
         time, Instrument('lyra'))
@@ -46,8 +46,9 @@ def test_can_handle_query(time):
     assert ans2 is False
 
 
-@settings(deadline=50000)
-@given(time_attr())
+@pytest.mark.parametrize("time", [
+    Time('2015/8/27', '2015/8/27'),
+    Time('2016/2/4', '2016/2/6')])
 @pytest.mark.remote_data
 def test_query(time):
     qr1 = LCClient.search(time, Instrument('lyra'))
