### 2019-11-05 a00e766
final fixes of tests to take into account scraping web
