### 2019-11-06 cb45111
Add an example of aligning HMI and AIA with reproject.
