### 2019-11-06 66785a2
Improve warnings squashing in wcs_4d thanks @astrofrog
