### 2019-11-06 3cf19f1
Update coordinates_in_asdf.py
