### 2019-11-06 15921a2
minor fix to maps.rst
