### 2019-11-08 39d8c68
Add an example showing fading between two maps
