### 2019-11-09 4f8b35a
removed pathlib and changed to bugfixes
