### 2019-11-09 856735c
Add some motivation to the top
