### 2019-11-10 660f1e7
Document some kwargs in mapsaving methods
