### 2019-11-11 #1832 @arfon
@mwcraig - this paper should work for now: [10.21105.joss.01832.pdf](https://github.com/openjournals/joss-reviews/files/3829309/10.21105.joss.01832.pdf)
