### 2019-11-11 #1832 @arfon
@Cadair - please fix up your paper authors and affiliations ASAP so that the reviewers can check it out. As it currently stands, the paper won't compile.
