### 2019-11-11 7e2f35c
Update joss_paper/paper.md
Co-Authored-By: Keith Hughitt <keith.hughitt@gmail.com>
