### 2019-11-11 6467e56
Add some Python 3.8 builds to the CI
