### 2019-11-11 032a6ff
Add some more comments throughout
