+++ examples/plotting/fading_between_maps.py
@@ -19,28 +19,36 @@ import sunpy.map
 from sunpy.data.sample import AIA_171_IMAGE, AIA_1600_IMAGE
 
 ###############################################################################
-# Start by loading two AIA maps from the sample data
+# Start by loading two AIA maps from the sample data.
+#
+# Note that these maps must be co-aligned. If they are taken at different times
+# or from different observation locations they will need to be aligned using
+# rotate() or reproject().
 map_171 = sunpy.map.Map(AIA_171_IMAGE)
 map_1600 = sunpy.map.Map(AIA_1600_IMAGE)
 
 ###############################################################################
 # First we create a figure, and add the axes that will show the maps. Then
-# plot both of the images on the same axes.
+# plot both of the images on the same axes. Finally, we add another axes that
+# contains the slider.
 fig = plt.figure()
+
+# Add the main axes. Note this is resized to leave room for the slider axes
 ax = fig.add_axes([0.1, 0.2, 0.9, 0.8], projection=map_171)
 
 im_1600 = map_1600.plot(axes=ax)
 im_171 = map_171.plot(axes=ax, alpha=0.5)
 ax.set_title('AIA 171 + 1600')
 
-###############################################################################
-# Now add another axes to the figure, which contains the slider
+# Add the slider axes
 ax_slider = fig.add_axes([0.25, 0.1, 0.65, 0.03])
 slider = Slider(ax_slider, 'Alpha', 0, 1, valinit=0.5)
 
+
 ###############################################################################
 # Finally, define what happens when the slider is changed and link this to the
-# slider
+# slider we set up above. In this case we just set the alpha (ie. transparencey)
+# of the 171 image.
 def update(val):
     alpha = slider.val
     im_171.set_alpha(alpha)
