### 2019-11-11 a7d566f
Update azure-pipelines.yml
