### 2019-11-11 86f7395
Add some missing doc requirements to conf.py
