### 2019-11-11 #1832 @mwcraig
> this paper should work for now: 10.21105.joss.01832.pdf

🤣 that second author, with the correct affiliation
