### 2019-11-11 a520f45
Update examples/plotting/fading_between_maps.py
Co-Authored-By: Stuart Mumford <stuart@cadair.com>
