### 2019-11-11 #1832 @mwcraig
Left some minor comments

+ On references at https://github.com/sunpy/sunpy/pull/3391#pullrequestreview-315211867
+ On "State of the field" at https://github.com/sunpy/sunpy/pull/3391#issuecomment-552664103

otherwise, this is good to go.
