### 2019-11-12 5fe932d
Add more affiliations from the main paper
