### 2019-11-12 977c5a1
Reduce sunpy.map import time
