### 2019-11-12 165a5e4
Syntax fixes in original yaml
