### 2019-11-12 2b81f61
reformat author list
