+++ joss_paper/paper.md
@@ -1,265 +1,357 @@
 ---
 title: 'SunPy: A Python package for Solar Physics'
 tags:
-  - Python
-  - Astronomy
-  - Solar physics
+- Python
+- Astronomy
+- Solar physics
 authors:
-  - name: Stuart J. Mumford
-    orcid: 0000-0003-4217-4642
-    affiliation:
-      - "SP2RC, School of Mathematics and Statistics, The University of Sheffield, UK"
-      - "Aperio Software Ltd., Headingley Enterprise and Arts Centre, Bennett Road, Leeds LS6 3HN"
-  - name: Nabil Freij
-    orcid: 0000-0002-6253-082X
-    affiliation: "Institute for Environmental Analytics, University of Reading, Reading RG6 6BX"
-  - name: Steven Christe
-    orcid: 0000-0001-6127-795X
-    affiliation: "NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA"
-  - name: Jack Ireland
-    orcid: 0000-0002-2019-8881
-    affiliation: "NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA"
-  - name: Florian Mayer
-  - name: V. Keith Hughitt
-    orcid: 0000-0003-0787-9559
-    affiliation: "Center for Cancer Research, National Cancer Institute, Bethesda, MD 20892-9760, USA"
-  - name: Albert Y. Shih
-    orcid: 0000-0001-6874-2594
-    affiliation: "NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA"
-  - name: Daniel F. Ryan
-    orcid: 0000-0001-8661-3825
-    affiliation:
-      - "American University, Washington, DC 20016, USA"
-      - "NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA"
-  - name: Simon Liedtke
-  - name: David Pérez-Suárez
-    orcid: 0000-0003-0784-6909
-    affiliation: "University College London, Gower Street, London, UK"
-  - name: Pritish Chakraborty
-    orcid: 0000-0001-8875-5819
-    affiliation: "Manav Rachna University"
-  - name: Vishnunarayan K I.
-  - name: Andrew Inglis
-    orcid: 0000-0003-0656-2437
-    affiliation: "Catholic University of America / NASA Goddard Space Flight Center, Greenbelt, MD 20771, USA"
-  - name: Punyaslok Pattnaik
-    orcid: 0000-0001-6216-2353
-    affiliation: 'International Institute of Information Technology, Hyderabad 500 032, India'
-  - name: Brigitta Sipőcz
-    orcid: 0000-0002-3713-6337
-    affiliation: "DIRAC Institute, Department of Astronomy, University of Washington, Seattle, WA 98195, USA"
-  - name: Rishabh Sharma
-  - name: Andrew Leonard
-    orcid: 0000-0001-5270-7487
-    affiliation: "Aperio Software Ltd., Headingley Enterprise and Arts Centre, Bennett Road, Leeds LS6 3HN"
-  - name: David Stansby
-    orcid: 0000-0002-1365-1908
-    affiliation: 'Mullard Space Science Laboratory, University College London, Surrey, UK'
-  - name: Russell Hewett
-    orcid: 0000-0001-8944-4705
-    affiliation: 'Department of Mathematics, Virginia Polytechnic Institute and State University, Blacksburg, VA 24061-0123, USA'
-  - name: Alex Hamilton
-  - name: Laura Hayes
[…]
