### 2019-11-12 #1832 @Cadair
:eyes: Do all the authors really require an affiliation? At least one has explicitly stated they don't have one they want to put on there. Also a lot of people contribute to OSS projects while explicitly not wearing their employers hat, so it might feel in appropriate to put affiliations down. @arfon Has this come up before?
