### 2019-11-12 66d5c17
Give everyone an affiliation
