+++ joss_paper/paper.md
@@ -7,347 +7,417 @@ tags:
 authors:
 - name: Stuart J. Mumford
   orcid: 0000-0003-4217-4642
-  affiliation: 1, 2
+  affiliation: 2, 3
 - name: Nabil Freij
   orcid: 0000-0002-6253-082X
-  affiliation: '3'
+  affiliation: '4'
 - name: Steven Christe
   orcid: 0000-0001-6127-795X
-  affiliation: '4'
+  affiliation: '5'
 - name: Jack Ireland
   orcid: 0000-0002-2019-8881
-  affiliation: '4'
+  affiliation: '5'
 - name: Florian Mayer
+  affiliation: '1'
 - name: V. Keith Hughitt
   orcid: 0000-0003-0787-9559
-  affiliation: '5'
+  affiliation: '6'
 - name: Albert Y. Shih
   orcid: 0000-0001-6874-2594
-  affiliation: '4'
+  affiliation: '5'
 - name: Daniel F. Ryan
   orcid: 0000-0001-8661-3825
-  affiliation: 6, 4
+  affiliation: 7, 5
 - name: Simon Liedtke
+  affiliation: '1'
 - name: David Pérez-Suárez
   orcid: 0000-0003-0784-6909
-  affiliation: '7'
+  affiliation: '8'
 - name: Pritish Chakraborty
   orcid: 0000-0001-8875-5819
-  affiliation: '8'
+  affiliation: '9'
 - name: Vishnunarayan K I.
+  affiliation: '1'
 - name: Andrew Inglis
   orcid: 0000-0003-0656-2437
-  affiliation: '9'
+  affiliation: '10'
 - name: Punyaslok Pattnaik
   orcid: 0000-0001-6216-2353
-  affiliation: '10'
+  affiliation: '11'
 - name: Brigitta Sipőcz
   orcid: 0000-0002-3713-6337
-  affiliation: '11'
+  affiliation: '12'
 - name: Rishabh Sharma
+  affiliation: '1'
 - name: Andrew Leonard
   orcid: 0000-0001-5270-7487
-  affiliation: '2'
+  affiliation: '3'
 - name: David Stansby
   orcid: 0000-0002-1365-1908
-  affiliation: '12'
+  affiliation: '13'
 - name: Russell Hewett
   orcid: 0000-0001-8944-4705
-  affiliation: '13'
+  affiliation: '14'
 - name: Alex Hamilton
+  affiliation: '1'
 - name: Laura Hayes
   orcid: 0000-0002-6835-2390
-  affiliation: '4'
+  affiliation: '5'
 - name: Asish Panda
+  affiliation: '1'
 - name: Matt Earnshaw
+  affiliation: '1'
 - name: Nitin Choudhary
   orcid: 0000-0001-6915-4583
-  affiliation: '14'
+  affiliation: '15'
 - name: Ankit Kumar
+  affiliation: '1'
 - name: Prateek Chanda
   orcid: 0000-0002-7068-2866
-  affiliation: '15'
-- name: Md Akramul Haque
   affiliation: '16'
+- name: Md Akramul Haque
+  affiliation: '17'
 - name: Michael S Kirk
   orcid: 0000-0001-9874-1429
-  affiliation: '9'
+  affiliation: '10'
 - name: Michael Mueller
+  affiliation: '1'
 - name: Sudarshan Konge
+  affiliation: '1'
 - name: Rajul Srivastava
+  affiliation: '1'
 - name: Yash Jain
   orcid: 0000-0001-5347-4734
-  affiliation: '17'
+  affiliation: '18'
 - name: Samuel Bennett
+  affiliation: '1'
 - name: Ankit Baruah
+  affiliation: '1'
[…]
