### 2019-11-12 #1832 @Cadair
Well I made it compile by doing that particular hack. My author list is being generated by a yaml translation script now, so it's easy enough to edit :grin:
