+++ joss_paper/author_processing.py
@@ -3,11 +3,6 @@ import io
 from ruamel.yaml import YAML
 yaml = YAML()
 
-with open("paper.md") as fd:
-    lines = fd.readlines()
-    start_line = lines.index("---\n")+1
-    end_line = lines.index("...\n")
-
 with open("metadata.yml") as fd:
     data = yaml.load(fd)
 
@@ -16,7 +11,7 @@ authors = data['authors']
 # Create a list of all affiliations and make the affiliation key a list (and always present)
 affiliations = []
 for auth in authors:
-    affil = auth.get('affiliation', "The SunPy Developers")
+    affil = auth.get('affiliation', None)
     if affil is not None:
         if not isinstance(affil, list):
             auth['affiliation'] = [affil]
@@ -25,30 +20,40 @@ for auth in authors:
 
 affiliations = set(affiliations)
 
-affiliation_indexes = {"The SunPy Developers": 1}
+affiliation_index = {"The SunPy Developers": 1}
 for auth in authors:
-    new_affils = []
+    new_affils = [1]
     if auth.get('affiliation') is not None:
         for affil in auth['affiliation']:
-            index = affiliation_indexes.get(affil, len(affiliation_indexes)+1)
-            affiliation_indexes[affil] = index
+            index = affiliation_index.get(affil, len(affiliation_index) + 1)
+            affiliation_index[affil] = index
             new_affils.append(index)
-        auth['affiliation'] = ', '.join(map(str, new_affils))
+    auth['affiliation'] = ', '.join(map(str, new_affils))
 
 
+# Build the affiliations section from the index
 affiliations = []
-for name, index in affiliation_indexes.items():
+for name, index in affiliation_index.items():
     affiliations.append({'name': name, 'index': index})
 
 data['affiliations'] = affiliations
 
 
+# Dump the new yaml to a string
 s = io.StringIO()
 yaml.default_flow_style = False
 yaml.width = 1000
 yaml.allow_unicode = True
 yaml.dump(data, s)
 s.seek(0)
+
+
+# Overwrite the paper markdown with the new metadata
+with open("paper.md") as fd:
+    lines = fd.readlines()
+
+start_line = lines.index("---\n") + 1
+end_line = lines.index("...\n")
 
 out_lines = lines[:start_line]
 out_lines += s.readlines()
