### 2019-11-12 a07cca7
Give all authors without an affiliation a "None" affiliation
