+++ joss_paper/paper.md
@@ -7,324 +7,324 @@ tags:
 authors:
 - name: Stuart J. Mumford
   orcid: 0000-0003-4217-4642
-  affiliation: 1, 2, 3
+  affiliation: 1, 2
 - name: Nabil Freij
   orcid: 0000-0002-6253-082X
-  affiliation: 1, 4
+  affiliation: '3'
 - name: Steven Christe
   orcid: 0000-0001-6127-795X
-  affiliation: 1, 5
+  affiliation: '4'
 - name: Jack Ireland
   orcid: 0000-0002-2019-8881
-  affiliation: 1, 5
+  affiliation: '4'
 - name: Florian Mayer
-  affiliation: '1'
+  affiliation: '5'
 - name: V. Keith Hughitt
   orcid: 0000-0003-0787-9559
-  affiliation: 1, 6
+  affiliation: '6'
 - name: Albert Y. Shih
   orcid: 0000-0001-6874-2594
-  affiliation: 1, 5
+  affiliation: '4'
 - name: Daniel F. Ryan
   orcid: 0000-0001-8661-3825
-  affiliation: 1, 7, 5
+  affiliation: 7, 4
 - name: Simon Liedtke
-  affiliation: '1'
+  affiliation: '5'
 - name: David Pérez-Suárez
   orcid: 0000-0003-0784-6909
-  affiliation: 1, 8
+  affiliation: '8'
 - name: Pritish Chakraborty
   orcid: 0000-0001-8875-5819
-  affiliation: 1, 9
+  affiliation: '9'
 - name: Vishnunarayan K I.
-  affiliation: '1'
+  affiliation: '5'
 - name: Andrew Inglis
   orcid: 0000-0003-0656-2437
-  affiliation: 1, 10
+  affiliation: '10'
 - name: Punyaslok Pattnaik
   orcid: 0000-0001-6216-2353
-  affiliation: 1, 11
+  affiliation: '11'
 - name: Brigitta Sipőcz
   orcid: 0000-0002-3713-6337
-  affiliation: 1, 12
+  affiliation: '12'
 - name: Rishabh Sharma
-  affiliation: '1'
+  affiliation: '5'
 - name: Andrew Leonard
   orcid: 0000-0001-5270-7487
-  affiliation: 1, 3
+  affiliation: '2'
 - name: David Stansby
   orcid: 0000-0002-1365-1908
-  affiliation: 1, 13
+  affiliation: '13'
 - name: Russell Hewett
   orcid: 0000-0001-8944-4705
-  affiliation: 1, 14
+  affiliation: '14'
 - name: Alex Hamilton
-  affiliation: '1'
+  affiliation: '5'
 - name: Laura Hayes
   orcid: 0000-0002-6835-2390
-  affiliation: 1, 5
+  affiliation: '4'
 - name: Asish Panda
-  affiliation: '1'
+  affiliation: '5'
 - name: Matt Earnshaw
-  affiliation: '1'
+  affiliation: '5'
 - name: Nitin Choudhary
   orcid: 0000-0001-6915-4583
-  affiliation: 1, 15
+  affiliation: '15'
 - name: Ankit Kumar
-  affiliation: '1'
+  affiliation: '5'
 - name: Prateek Chanda
   orcid: 0000-0002-7068-2866
-  affiliation: 1, 16
+  affiliation: '16'
 - name: Md Akramul Haque
-  affiliation: 1, 17
+  affiliation: '17'
 - name: Michael S Kirk
   orcid: 0000-0001-9874-1429
-  affiliation: 1, 10
+  affiliation: '10'
 - name: Michael Mueller
-  affiliation: '1'
+  affiliation: '5'
 - name: Sudarshan Konge
-  affiliation: '1'
+  affiliation: '5'
[…]
