### 2019-11-12 ac1c80e
Fix multiple places where we assumed the type of header items
