+++ sunpy/map/sources/suvi.py
@@ -104,6 +104,6 @@ class SUVIMap(GenericMap):
     @classmethod
     def is_datasource_for(cls, data, header, **kwargs):
         """Determines if header corresponds to an AIA image"""
-        return header.get("instrume", "").startswith(
+        return str(header.get("instrume", "")).startswith(
             "GOES-R Series Solar Ultraviolet Imager"
         )
