### 2019-11-12 2e793ca
Fix multiple places where we assumed the type of header items
