### 2019-11-12 d99a19b
Fix multiple places where we assumed the type of header items
