### 2019-11-12 96104b8
Fix multiple places where we assumed the type of header items
