+++ sunpy/map/sources/iris.py
@@ -48,7 +48,7 @@ class SJIMap(GenericMap):
     @classmethod
     def is_datasource_for(cls, data, header, **kwargs):
         """Determines if header corresponds to an IRIS SJI image"""
-        tele = header.get('TELESCOP', '').startswith('IRIS')
-        obs = header.get('INSTRUME', '').startswith('SJI')
+        tele = str(header.get('TELESCOP', '')).startswith('IRIS')
+        obs = str(header.get('INSTRUME', '')).startswith('SJI')
         level = header.get('lvl_num') == 1
         return tele and obs
