### 2019-11-12 e6ea505
Construct the WCS from the fits header
