### 2019-11-12 5d2b31d
Fix darkest VSO errors found while trying to query too much WISPR data
