### 2019-11-13 #1832 @mwcraig
@arfon -- I made a comment on the [sunpy JOSS paper](https://github.com/sunpy/sunpy/pull/3391#issuecomment-552664103) that they hadn't addressed the state of the field in the JOSS paper.  @mbobra mentioned that it is addressed in the accompanying ApJ paper:

> We address this in the introduction section of the accompanying ApJ Paper. Does this satisfy the review criteria or would you like us to explicitly address this in the JOSS paper?

I've read that section of the ApJ paper and it is very good and probably too long for the JOSS paper. Does it need to be addressed in the JOSS paper? Would a reference to the ApJ paper suffice?

My leaning would be towards a reference in the JOSS paper to the discussion in the ApJ paper but I haven't reviewed a JOSS/ApJ combination before.
