### 2019-11-13 0b16249
Apply suggestions from code review
Co-Authored-By: Stuart Mumford <stuart@cadair.com>
