### 2019-11-13 fe089fa
Don't make the Type and ABC to fix metaclass conflict
