+++ sunpy/util/types.py
@@ -1,10 +1,10 @@
 """
 A collection of ABCs to use to shortcut having to import things.
 """
-import abc
 
 
-class DatabaseEntryType(metaclass=abc.ABCMeta):
+class DatabaseEntryType:
     """
     A base class for `sunpy.database.tables.DatabaseEntry`.
+    This class should not be used directly.
     """
