### 2019-11-13 b5269d7
Update color_tables.py
