### 2019-11-13 905c691
Made all "color_tables" functions take units only.
