+++ setup.cfg
@@ -37,7 +37,7 @@ install_requires =
   matplotlib>=1.3
   pandas
   astropy>=3.1
-  parfive[ftp]
+  parfive[ftp]>=1.1.0rc1
   importlib_resources;python_version<"3.8"
 
 [options.extras_require]
