### 2019-11-13 9071446
Update color_tables.py
