### 2019-11-13 b062449
Update color_tables.py
