### 2019-11-13 #1832 @arfon
> I've read that section of the ApJ paper and it is very good and probably too long for the JOSS paper. Does it need to be addressed in the JOSS paper? Would a reference to the ApJ paper suffice?

I think this is OK, we just need to make sure it's cited properly in the body of the text.
