### 2019-11-13 1a699de
Add support for tick settings in ArrayAnimatorWCS.
