+++ sunpy/visualization/animator/tests/test_wcs.py
@@ -143,6 +143,7 @@ def test_array_animator_wcs_coord_params(wcs_4d):
             'format_unit': u.deg,
             'major_formatter': 'hh:mm:ss',
             'axislabel': 'Longitude',
+            'ticks': {'spacing': 10*u.arcsec}
         }
     }
 
