### 2019-11-13 cefab42
replace ImageAnimatorWCS example with a ArrayAnimatorWCS example
