### 2019-11-13 b574670
Update color_tables.py
