+++ changelog/3495.feature.rst
@@ -1,2 +1,2 @@
 Made a couple of module imports lazy to reduce the import time of sunpy.map by
-~20%.
+~40%.
