### 2019-11-13 2345f24
Add a test for array obstimes
