### 2019-11-13 7ea5a1b
addresses https://github.com/sunpy/sunpy/pull/3391#issuecomment-552664103
