### 2019-11-14 #1832 @mwcraig
Aside from a couple of cosmetic issues with the new references (commented on the paper PR) this is ready to go. The reference to the ApJ paper isn't complete yet, but I assume that will be worked out. Thanks for the quick responses to the review comments @Cadair and @mbobra!
