### 2019-11-15 #1832 @ceb8
Aside from the incomplete ApJ paper reference issue noted by @mwcraig this is ready.
