### 2019-11-15 #1832 @arfon
Thanks @ceb8 and @mwcraig.

@Cadair - I don't think we can finalize the ApJ citation until we have a DOI for this paper. @crawfordsm - do we have a DOI yet for the accompanying ApJ paper?
