### 2019-11-19 #1832 @Cadair
@arfon We would like to document how we have sorted our author list somewhere, in the name of transparency and clarity. Do you have any suggestions about where we should put that information?
