### 2019-11-21 #1832 @arfon
> @arfon We would like to document how we have sorted our author list somewhere, in the name of transparency and clarity. Do you have any suggestions about where we should put that information?

Something like this will allow you to have a footnote in the paper: https://github.com/sunpy/sunpy/compare/joss_paper...arfon:patch-1
