### 2019-12-05 #1832 @danielskatz
👋 @arfon - it looks like this is waiting for the ApJ paper's DOI, correct?  Assuming so, I'm going to mark it as paused for now.
