### 2020-01-07 #1832 @danielskatz
👋 Hi all - just checking that we are still waiting for the ApJ paper's DOI?
