### 2020-01-07 #1832 @arfon
> 👋 Hi all - just checking that we are still waiting for the ApJ paper's DOI?

Yes, I believe so. Any news on this front @crawfordsm ?
