### 2020-01-07 #1832 @Cadair
As far as I know we have seen some proofs of the ApJ paper, but not a DOI yet.
