### 2020-02-06 #1832 @arfon
@Cadair - apparently the AAS DOI is `10.3847/1538-4357/ab4f7a`. I've opened a PR adding this in https://github.com/sunpy/sunpy/pull/3771
