### 2020-02-06 #1832 @Cadair
Thanks @arfon what is the procedure now? I need to make a release with the paper in and the author list on Zenodo needs to match the paper?
