### 2020-02-06 #1832 @arfon
>  I need to make a release with the paper in and the author list on Zenodo needs to match the paper?

Ideally yes.
