### 2020-02-06 #1832 @Cadair
Ok I need to write a script to generate the zenodo metadata from the paper metadata. I will try and get that sorted in the next couple of days.
