### 2020-02-07 #1832 @Cadair
@arfon other than the author list breaking the sidebar, it also seems the footer with the author names in etc has got the `[` from the footnote syntax after my name?
