### 2020-02-07 #1832 @crawfordsm
In case you still need it, the ApJ Doi is:

DOI: 10.3847/1538-4357/ab4f7a
