### 2020-02-07 #1832 @arfon
@Cadair - this PR should fix the footnote: https://github.com/sunpy/sunpy/pull/3774
