### 2020-02-07 #1832 @Cadair
@crawfordsm or @arfon the bibtex for the AAS reference is also incomplete. Now we have a DOI do we have access to the full reference bibtex?
