### 2020-02-12 #1832 @arfon
@Cadair - looks like the AAS paper is published now, can you figure out the details from this: https://iopscience.iop.org/article/10.3847/1538-4357/ab4f7a ?
