### 2020-02-12 #1832 @arfon
@Cadair because I'm feeling nice today 😸 https://github.com/sunpy/sunpy/compare/master...arfon:patch-4
