### 2020-02-12 #1832 @Cadair
Thanks! The ApJ site was giving me a 500 when I tried to download the bibtex :angry:
