### 2020-02-12 #1832 @arfon
@Cadair - does [this Zenodo release](https://zenodo.org/record/3604353#.XkRABhNKind) include all of the changes that resulted from this JOSS review?
