### 2020-02-12 #1832 @Cadair
@arfon Probably. I was planning on doing a 1.0.8 with the correct author info etc shortly.
