### 2020-02-13 #1832 @Cadair
@arfon 1.0.8 has been released the zenodo record is here: https://zenodo.org/record/3666625
