### 2020-02-13 #1832 @arfon
@Cadair - please reformat the YAML header of your `paper.md` to match the [documented example](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).

Specifically, your `affiliations` block is currently empty. Affiliations in JOSS papers need to be described separately and then referenced by their index which allows for more reuse of affiliations between authors.
