### 2020-02-14 #1832 @Cadair
@arfon What version of the `paper.md` is it pulling? The version on master and on the 1.0 both have a filled in `affiliations` block with references to the indexes etc. I think everything is spelt correctly and nothing is missing. :confused:
