### 2020-02-14 #1832 @arfon
BTW @Cadair, the formatting of the PDF (margin on second page) is as good as I could get the layout without completely rebuilding our LaTeX template. I hope this is acceptable :-)
